Split into several crates to have the parsing into a separate crate.
