+++ parsers/Cargo.toml
@@ -1,0 +1,20 @@
+
+[package]
+name = "uil_parsers"
+version = "0.0.1"
+authors = ["Nemikolh <joan.varvenne@gmail.com>"]
+homepage = "http://uil-lang.github.io/"
+repository = "https://github.com/uil-lang/uil-rs"
+license = "MIT"
+keywords = ["ui", "parser"]
+description = "Parsers for uil."
+
+[dependencies]
+xml-rs = "*"
+phf = "*"
+phf_macros = "*"
+num = "*"
+
+[dependencies.uil_shared]
+path = "../shared"
+version = "=0.0.1"
