+++ parsers/src/deps/mod.rs
@@ -2,12 +2,13 @@
 mod parser;
 
 // Dependencies
-use std::collections::HashMap;
 use std::io::{BufRead, BufReader};
-use std::path::{PathBuf, Path};
+use std::path::Path;
 use std::fs::File;
 
 use report::ErrorReporter;
+use uil_shared::deps::StyleDefinitions;
+
 
 /// Convenient function to parse a style from a BufRead.
 ///
@@ -56,36 +57,4 @@ pub fn parse_file<E, P>(reporter: E, path: P) -> StyleDefinitions
         path.as_ref().parent().unwrap_or(Path::new(".")).to_path_buf()
     );
     parser.parse()
-}
-
-pub struct StyleDefinitions {
-    pub defs: HashMap<String, Constructor>,
-}
-
-impl StyleDefinitions {
-    pub fn new() -> StyleDefinitions {
-        StyleDefinitions {
-            defs: HashMap::new(),
-        }
-    }
-
-    pub fn insert(&mut self, key: String, ctor: Constructor) {
-        self.defs.insert(key, ctor);
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum Constructor {
-    /// None type -> Constructor failed loading the resource.
-    None,
-    /// Number [0-9]+
-    Number(f32),
-    /// String ".+"
-    Quote(String),
-    /// Font(path, width, height)
-    Font(String, f32, f32),
-    /// TODO: replace String by the type Path
-    /// Image(path, width, height, offset-x, offset-y)
-    Image(PathBuf, Option<f32>, Option<f32>, Option<f32>, Option<f32>),
-    // Add other construtor here...
 }
