+++ shared/Cargo.toml
@@ -1,0 +1,10 @@
+
+[package]
+name = "uil_shared"
+version = "0.0.1"
+authors = ["Nemikolh <joan.varvenne@gmail.com>"]
+homepage = "http://uil-lang.github.io/"
+repository = "https://github.com/uil-lang/uil-rs"
+license = "MIT"
+keywords = ["ui", "parser"]
+description = "Support code shared by uil libraries."
