+++ shared/src/lib.rs
@@ -1,0 +1,7 @@
+#![feature(core)]
+
+pub mod asset;
+pub mod resource;
+pub mod style;
+pub mod deps;
+pub mod markup;
