+++ src/resource.rs
@@ -2,21 +2,21 @@ use std::path::Path;
 use glium::{self, Display};
 use glium::texture::CompressedTexture2d;
 use image::{self, GenericImage};
+use uil_shared::resource::new_resource_id;
+
+
+// Reexport
+
+pub use uil_shared::resource::ResourceId;
+pub use uil_shared::resource::BasicResourceManager;
 
 // ======================================== //
 //                INTERFACE                 //
 // ======================================== //
 
-#[derive(Copy, Clone)]
-pub struct ResourceId(usize);
-
-pub trait ResourceManager {
-
-    fn get_texture_id(&mut self, p: &Path) -> ResourceId;
+pub trait ResourceManager: BasicResourceManager {
 
     fn get_texture(&self, id: ResourceId) -> &CompressedTexture2d;
-
-    fn get_image_dimensions(&self, id: ResourceId) -> (u32, u32);
 }
 
 pub fn create_resource_manager<'a>(display: &'a Display) -> ResourceManagerImpl {
@@ -52,7 +52,7 @@ impl<'a> ResourceManagerImpl<'a> {
     }
 }
 
-impl<'a> ResourceManager for ResourceManagerImpl<'a> {
+impl<'a> BasicResourceManager for ResourceManagerImpl<'a> {
 
     fn get_texture_id(&mut self, p: &Path)
         -> ResourceId
@@ -66,13 +66,7 @@ impl<'a> ResourceManager for ResourceManagerImpl<'a> {
             img_width: iw,
             img_height: ih
         });
-        ResourceId(id)
-    }
-
-    fn get_texture(&self, id: ResourceId)
-        -> &CompressedTexture2d
-    {
-        &self.textures[id.0].handle
+        unsafe { new_resource_id(id) }
     }
 
     fn get_image_dimensions(
@@ -80,32 +74,42 @@ impl<'a> ResourceManager for ResourceManagerImpl<'a> {
         id: ResourceId)
         -> (u32, u32)
     {
-        (self.textures[id.0].img_width, self.textures[id.0].img_height)
+        unsafe {
+            (self.textures[id.get()].img_width, self.textures[id.get()].img_height)
+        }
+    }
+}
+
+impl<'a> ResourceManager for ResourceManagerImpl<'a> {
+
+    fn get_texture(&self, id: ResourceId)
+        -> &CompressedTexture2d
+    {
+        unsafe { &self.textures[id.get()].handle }
     }
 }
 
 pub struct NullResourceManager;
 
-impl ResourceManager for NullResourceManager {
+impl BasicResourceManager for NullResourceManager {
 
     fn get_texture_id(&mut self, _: &Path)
         -> ResourceId
     {
-        ResourceId(0)
+        unsafe { new_resource_id(0) }
     }
+
+    fn get_image_dimensions(&self, _: ResourceId) -> (u32, u32) {
+        (0, 0)
+    }
+}
+
+impl ResourceManager for NullResourceManager {
 
     fn get_texture(&self, _: ResourceId)
         -> &CompressedTexture2d
     {
         panic!("NullResourceManager purpose is for test only,\
                 it has a limited use.");
-    }
-
-    fn get_image_dimensions(
-        &self,
-        _: ResourceId)
-        -> (u32, u32)
-    {
-        (0, 0)
     }
 }
